rest/requirements.txt
Django==1.11.4
django-filter==1.1.0
djangorestframework==3.6.3
drf-url-filters==0.5.1
mysqlclient==1.3.10
pytz==2017.2
voluptuous==0.10.5
